Move proof.go file to a separate package
